first implementation of logic plan generation

+++ src/main/java/edu/ucsd/cs/triton/expression/FloatExpression.java
@@ -3,7 +3,7 @@ package edu.ucsd.cs.triton.expression;
 public class FloatExpression extends BaseExpression {
 	private float _value;
 	
-	public FloatExpression(int value) {
+	public FloatExpression(float value) {
 		super();
 		_value = value;
 		_isLeaf = true;
